[FileIO] Move GmshReader to FileIO.
