+++ Applications/DataExplorer/mainwindow.cpp
@@ -64,9 +64,9 @@
 #include "Applications/FileIO/FEFLOW/FEFLOWMeshInterface.h"
 #include "MeshLib/IO/Legacy/MeshIO.h"
 #include "MeshLib/IO/readMeshFromFile.h"
-#include "MeshLib/IO/GmshReader.h"
 #include "Applications/FileIO/AsciiRasterInterface.h"
 #include "Applications/FileIO/PetrelInterface.h"
+#include "Applications/FileIO/Gmsh/GmshReader.h"
 #include "Applications/FileIO/TetGenInterface.h"
 #include "Applications/FileIO/XmlIO/Qt/XmlGspInterface.h"
 #include "GeoLib/IO/GMSHInterface.h"
