+++ Applications/FileIO/CMakeLists.txt
@@ -13,6 +13,8 @@ if(QT4_FOUND)
 endif()
 GET_SOURCE_FILES(SOURCES_RAPID_XML XmlIO/Rapid)
 set(SOURCES ${SOURCES} ${SOURCES_RAPID_XML})
+GET_SOURCE_FILES(SOURCES_GMSH Gmsh)
+set(SOURCES ${SOURCES} ${SOURCES_GMSH})
 
 if(QT4_FOUND)
     GET_SOURCE_FILES(SOURCES_FEFLOW FEFLOW)
